pyproject.toml
[project]
name = "flow-unrolled"
version = "0.0.0"
requires-python = "==3.11.15"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.5",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.13.0",
    "attrs==26.1.0",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.7",
    "click==8.3.1",
    "einops==0.8.2",
    "fastmri==0.3.0",
    "filelock==3.25.2",
    "frozenlist==1.8.0",
    "fsspec==2026.3.0",
    "gitdb==4.0.12",
    "GitPython==3.1.46",
    "h11==0.16.0",
    "h5py==3.16.0",
    "hf-xet==1.4.3",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface_hub==1.8.0",
    "idna==3.11",
    "ImageIO==2.37.3",
    "Jinja2==3.1.6",
    "lazy-loader==0.5",
    "lightning-utilities==0.15.3",
    "markdown-it-py==4.0.0",
    "MarkupSafe==3.0.3",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.7.1",
    "networkx==3.6.1",
    "numpy==2.4.4",
    "nvidia-ml-py==13.590.48",
    "nvitop==1.6.2",
    "packaging==26.0",
    "pandas==3.0.2",
    "pillow==12.2.0",
    "platformdirs==4.9.4",
    "propcache==0.4.1",
    "protobuf==6.33.6",
    "psutil==7.2.2",
    "ptflops==0.7.5",
    "pydantic==2.12.5",
    "pydantic_core==2.41.5",
    "Pygments==2.20.0",
    "python-dateutil==2.9.0.post0",
    "pytorch-lightning==2.6.1",
    "PyYAML==6.0.3",
    "requests==2.33.1",
    "rich==14.3.3",
    "runstats==2.0.0",
    "safetensors==0.7.0",
    "scikit-image==0.26.0",
    "scipy==1.17.1",
    "sentry-sdk==2.57.0",
    "setuptools==81.0.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.3",
    "sympy==1.14.0",
    "tensorboard==2.20.0",
    "tensorboardx>=2.6.5",
    "tifffile==2026.3.3",
    "timm==1.0.26",
    "torch==2.7.1",
    "torchaudio==2.7.1",
    "torchmetrics==1.9.0",
    "torchvision==0.22.1",
    "tqdm==4.67.3",
    "typer==0.24.1",
    "typing-inspection==0.4.2",
    "typing_extensions==4.15.0",
    "urllib3==2.6.3",
    "wandb==0.25.1",
    "yarl==1.23.0",
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"

[tool.uv.sources]
torch = { index = "pytorch-cu126" }
torchvision = { index = "pytorch-cu126" }
torchaudio = { index = "pytorch-cu126" }

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true
